fix: map class names by mojang mapping.

+++ src/main/kotlin/com/github/ordom/remapper/transformer/BytecodeRemapper.kt
@@ -1,5 +1,6 @@
 package com.github.ordom.remapper.transformer
 
+import com.github.ordom.remapper.CLASS_MAPPED_SEARGE
 import com.github.ordom.remapper.INTERMEDIARY
 import com.github.ordom.remapper.SEARGE
 import com.github.ordom.remapper.VERSION
@@ -10,6 +11,8 @@ import com.github.ordom.remapper.metadata.ForgeMetadata
 import com.github.ordom.remapper.metadata.ForgePackMeta
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.Serializable
+import kotlinx.serialization.decodeFromString
+import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import net.fabricmc.loom.util.TinyRemapperMappingsHelper
 import net.fabricmc.tinyremapper.NonClassCopyMode
@@ -18,11 +21,14 @@ import net.fabricmc.tinyremapper.TinyRemapper
 import org.objectweb.asm.ClassWriter
 import org.objectweb.asm.Opcodes
 import java.nio.file.Path
+import java.util.jar.Attributes
+import java.util.jar.Manifest
 import java.util.zip.ZipEntry
 import java.util.zip.ZipOutputStream
 import kotlin.io.path.*
 
 private typealias strM<T> = Map<String, T>
+val MIXIN_CONFIGS = Attributes.Name("MixinConfigs")
 
 @Serializable
 private data class FabricRefMap(
@@ -42,7 +48,7 @@ class BytecodeRemapper(
 ) {
     private val mapping = i2s.merge()
     val remapper: TinyRemapper = TinyRemapper.newRemapper()
-        .withMappings(TinyRemapperMappingsHelper.create(mapping, INTERMEDIARY, SEARGE, false))
+        .withMappings(TinyRemapperMappingsHelper.create(mapping, INTERMEDIARY, CLASS_MAPPED_SEARGE, false))
         .build()
 
     @OptIn(ExperimentalPathApi::class)
@@ -54,19 +60,20 @@ class BytecodeRemapper(
         val translator = SignatureTranslator(mapping)
         val output = OutputConsumerPath.Builder(path)
             .build()
-        remapper.readInputs(jarFile)
-        remapper.apply(output)
+        val inputTag = remapper.createInputTag()
+        remapper.readInputs(inputTag, jarFile)
+        remapper.apply(output, inputTag)
 
         output.addNonClassFiles(jarFile, NonClassCopyMode.FIX_META_INF, remapper)
+        val fabricMetadata = FabricMetadata.fromJson(path / "fabric.mod.json")
         // remap mixins classes
-        val mixinConfigs = FabricMetadata.fromJson(path / "fabric.mod.json").mixins
-        mixinConfigs.asSequence()
+        fabricMetadata.mixins.asSequence()
             .map { FabricMetadata.MixinConfig.fromJson(path / it) }
             .map { (path / it.refMap).toFile() }
             .map { it to it.readText() }
-            .mapValues { JSON.decodeFromString(FabricRefMap.serializer(), it) }
+            .mapValues { JSON.decodeFromString<FabricRefMap>(it) }
             .mapValues { mapToSrg(it, translator) }
-            .mapValues { JSON.encodeToString(FabricRefMap.serializer(), it) }
+            .mapValues { JSON.encodeToString(it) }
             .forEach { it.first.writeText(it.second) }
 
         // generate forge meta
@@ -74,12 +81,13 @@ class BytecodeRemapper(
         path / "pack.mcmeta" writeText forgePackMeta.toString()
         // generate forge toml
         val forgeMetadata = ForgeMetadata(
-            modid = "ordom_forge",
-            version = "0.1",
-            name = "OrdomForge",
-            description = "Generated from fabric",
-            license = "LGPL-2.1",
-        ) // todo: parse from fabric.mod.json
+            // '-' is allowed in fabric mod id, but not in forge
+            modid = fabricMetadata.id.replace('-', '_'),
+            version = fabricMetadata.version,
+            name = fabricMetadata.name,
+            description = fabricMetadata.description ?: "",
+            license = fabricMetadata.license ?: "All rights reserved (Unlicensed)",
+        )
         path / "META-INF" / "mods.toml" writeText """
             modLoader = "javafml"
             loaderVersion = "[45,)"
@@ -92,38 +100,39 @@ class BytecodeRemapper(
         """.trimIndent()
 
         // aad info to manifest
-        val manifest = (path / "META-INF" / "MANIFEST.MF").toFile()
-        manifest.appendText("""
-            MixinConfigs: ${mixinConfigs.joinToString(",")}
-            Specification-Title: ${forgeMetadata.modid}
-            Specification-Version: 1
-            Implementation-Title: ${forgeMetadata.name}
-            Implementation-Version: ${forgeMetadata.version}
-            Mod-Translator: Ordom
-            X-Ordomal-Remapped: true
-            X-Ordomal-Origin-Mod: ${jarFile.name}
-            X-Ordomal-Remapped-Mod: ${jarFile.nameWithoutExtension}-remapped.jar
-            X-Ordomal-Remapped-Mod-Id: ordom_forge
-            X-Ordomal-Remapped-Mod-Version: 0.1
-            X-Ordomal-Version: $VERSION
-        """.trimIndent())
+        val manifest = Manifest((path / "META-INF" / "MANIFEST.MF").inputStream())
+        manifest.mainAttributes[MIXIN_CONFIGS] = fabricMetadata.mixins.joinToString(",")
+        manifest.mainAttributes[Attributes.Name.SPECIFICATION_TITLE] = forgeMetadata.modid
+        manifest.mainAttributes[Attributes.Name.SPECIFICATION_VERSION] = "1"
+        manifest.mainAttributes[Attributes.Name.IMPLEMENTATION_TITLE] = forgeMetadata.name
+        manifest.mainAttributes[Attributes.Name.IMPLEMENTATION_VERSION] = forgeMetadata.version
+        manifest.mainAttributes[Attributes.Name("Mod-Translator")] = "Ordom"
+        manifest.mainAttributes[Attributes.Name("X-Ordomal-Remapped")] = "true"
+        manifest.mainAttributes[Attributes.Name("X-Ordomal-Origin-Mod")] = jarFile.name
+        manifest.mainAttributes[Attributes.Name("X-Ordomal-Version")] = VERSION
+        manifest.write((path / "META-INF" / "MANIFEST.MF").outputStream())
 
         // generate forge mod class
-        val asm = ClassWriter(0)
-        asm.visit(61, Opcodes.ACC_PUBLIC, "ordom/remapped/ordom_forge", null, "java/lang/Object", null)
-        asm.visitAnnotation("Lnet/minecraftforge/fml/common/Mod;", true).visit("value", "ordom_forge")
-        asm.newMethod("ordom/remapped/ordom_forge", "<init>", "()V", false)
-        asm.visitMethod(Opcodes.ACC_PUBLIC, "<init>", "()V", null, null).apply {
-            visitCode()
-            visitVarInsn(Opcodes.ALOAD, 0)
-            visitMethodInsn(Opcodes.INVOKESPECIAL, "java/lang/Object", "<init>", "()V", false)
-            visitInsn(Opcodes.RETURN)
-            visitMaxs(1, 1)
+        val asm = ClassWriter(0).apply {
+            val classSig = "ordom/remapped/${forgeMetadata.modid}"
+            visit(61, Opcodes.ACC_PUBLIC, classSig, null, "java/lang/Object", null)
+            visitAnnotation("Lnet/minecraftforge/fml/common/Mod;", true).visit("value", forgeMetadata.modid)
+            newMethod(classSig, "<init>", "()V", false)
+            visitMethod(Opcodes.ACC_PUBLIC, "<init>", "()V", null, null).apply {
+                visitCode()
+                visitVarInsn(Opcodes.ALOAD, 0)
+                visitMethodInsn(Opcodes.INVOKESPECIAL, "java/lang/Object", "<init>", "()V", false)
+                visitInsn(Opcodes.RETURN)
+                visitMaxs(1, 1)
+                visitEnd()
+            }
             visitEnd()
         }
-        asm.visitEnd()
         (path / "ordom" / "remapped").createDirectories()
-        (path / "ordom" / "remapped" / "ordom_forge.class").writeBytes(asm.toByteArray())
+        (path / "ordom" / "remapped" / "${forgeMetadata.modid}.class").writeBytes(asm.toByteArray())
+
+        // delete fabric meta
+        (path / "fabric.mod.json").deleteIfExists()
 
         // pack into jar file
         val outputFile = path.resolveSibling("${path.name}.jar").toFile()
@@ -136,9 +145,7 @@ class BytecodeRemapper(
                 zipOutputStream.closeEntry()
             }
         }
-        if (deleteTmp) {
-            path.deleteRecursively()
-        }
+        if (deleteTmp) path.deleteRecursively()
         return path
     }
 
@@ -151,16 +158,16 @@ class BytecodeRemapper(
     private fun mapToSrg(mixinRefMap: FabricRefMap, translator: SignatureTranslator): FabricRefMap {
         val mappings = mixinRefMap.mappings?.map { (mixinClass, map) ->
             mixinClass to map.map { (mixinRef, inter) ->
-                val newInter = translator.translate(inter, SEARGE)
+                val newInter = translator.translate(inter, CLASS_MAPPED_SEARGE)
                 mixinRef to newInter
             }.toMap()
         }?.toMap()
         val data = mixinRefMap.data?.map { (namespace, map1) ->
             if (namespace != "named:intermediary")
                 error("Unknown namespace key: $namespace")
-            SEARGE to map1.map { (mixinClass, map2) ->
+            namespace to map1.map { (mixinClass, map2) ->
                 val newMap = map2.map { (mixinRef, inter) ->
-                    mixinRef to translator.translate(inter, SEARGE)
+                    mixinRef to translator.translate(inter, CLASS_MAPPED_SEARGE)
                 }.toMap()
                 mixinClass to newMap
             }.toMap()
